requirements.txt
django
python-docx
spacy
nltk
requests
pillow
python-pptx
PyPDF2
# Add runtime requirements for deployment
gunicorn
whitenoise
